WIP: `CatalogThemes.vue` → `App.vue`
